Literature Review Commit
Here is the commit of the completed literature review, and the articles that I used documented as references following harvard style referencing. The literature review covers literature that discusses many topics related to the subject of the thesis. Firstly how people have been hooked to electronics and games since their inception, and how the use of these electronics in education if implemented well could improve the teaching of difficult topics to students by far. Next I reviewed literature that go into detail with tests to determine the extent that children understand fractions having completed the topic at school going through the traditional methods of teaching and learning.
Next I went over literature that talks about the problems there are with how fractions are being taught today.
Next I reviewed literature on the effectiveness of gamification and how this could possibly improve the learning experience for young students making their life easier and that of the teacher too, as it is hard to teach difficult concepts to a class of students who are not understanding you. I continue by reviewing more literature on how to improve the gamification experience and link this with teaching fractions.