Create Methodology.pdf
In this commit i uploaded my completed methodology.  In this section of my thesis I have 4 sub sections. the first is called "Description of the Game" where i describe everything there is in the game from the different menu screens to the mechanics and what happens during the levels. In the second section I describe how the users interact with the came. Since the game is intended for young children I created a simple UI where all the children need to do is press a button with the correct answer.  I also showed screen shots to help the descriptions I provide. In the third section I talked about the participants who took part in the sessions where they played the game in class. For the last secion I talked about what was organised by me and the teachers to get the children to play the game, and what was carried out during the sessions