Create RsultsAndDiscussion.pdf
In this commit I uploaded the Results and discussion section of my thesis. In this section I start off by talking about why i collected the quantitative and qualitative data that I collected and how I did so. Then I talked about the use of the data. I interviewed the teacher that over saw the sessions with the children and asked her about what she thinks about the potential of gamification in education after having experienced these sessions. Mostly positive feedback was recieved. Next I talked about the Performance of the Prototype game that was used in the sessions, and how in order to create the game a specific method of gamification was followed. Next I summurized each session that took place at DeLa Salle where the game was played. I talked about the behaviours of the children and interesting things that were noted. And finally I answered my research questions to conclude this section, I backed up the answeres to my research questions with research papers done by other people.