Create Abstract&Introduction&Conc.pdf
In this commit I uploaded my abstract, introduction and conclusion. In the abstract I sumarrised the reason behind why I wrote my thesis (the problem that exists), I also summarised what I did for my methodology followed by a summary of the results I obtained.
In the introduction I elaborated more on the problem that exists and quoted some articles that discuss this problem.  I continue with talking about the motivation I had behind writing this thesis, the research quesitons I had, and I gave a brief summary of the following chapters of my thesis.
For the conclusion I discussed the results I obtained and backed up my results with those obtained in other research papers